Add UMD support using rollup
