+++ .gitignore
@@ -1,3 +1,4 @@
 *~
 .*.sw?
 node_modules
+lib
